Added depth of model as a hyperparameter.
The sizes of hidden layers are chosen based on the number of layers, ORIGINAL_DIM and the size of a latent vector. The range [latent vector's size, ORIGINAL_DIM] is equally split into the number of layers chunks. Then size of each layer is independently sampled.

Rewrote the model. Split up VAE class which is now a keras.Model and created VAEHandler responsible for manufacturing the model and management of training process.