Distributed training.
Now both the training and the hyperparameter optimization may be performed on multiple GPUs within one machine. The strategy which is used for a distributed training is MirroredStrategy. Also implemented GPULimiter which allows to specify which GPUs should be used during the training and what is the maximum memory limit that TF can allocate for a training.